# Optimize MetricData for the case where labels are accessed