Reduce the number of total allocations and the amount of allocate bytes:
1) Remove unnecessary Label object.
2) Remove unnecessary internal pimpl when only the labels internal is needed. This change reduces with 2 allocations.
3) Remove sliceChanged variable and directly change the `orig` pointer when updating a slice.

Before:
```
$go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              290311723                4.03 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                1199534               970 ns/op            1216 B/op         18 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               1233117               967 ns/op            1216 B/op         18 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            1000000              1223 ns/op            1392 B/op         20 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              1000000              1011 ns/op            1280 B/op         19 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 10.458s
```

After:
```
go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              317034267                3.64 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                1399080               849 ns/op            1088 B/op         16 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               1414124               846 ns/op            1088 B/op         16 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            1383159               862 ns/op            1152 B/op         16 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              1377898               935 ns/op            1152 B/op         16 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 10.045s
```